013-plant-disease-predictor-master/plant-disease-predictor-master/requirements.txt
PySimpleGUI==4.41.2
h5py==2.10.0
tensorflow==1.5.0
Keras==2.2.4
Pillow==8.1.1
